added homework alarm

+++ app/src/main/java/uz/suhrob/darsjadvalitatuuf/models/HomeWork.kt
@@ -3,4 +3,4 @@ package uz.suhrob.darsjadvalitatuuf.models
 /**
  * Created by User on 17.03.2020.
  */
-data class HomeWork(val id: Int, val title: String, val content: String, val weekDay: WeekDay, val order: Int)
+data class HomeWork(val id: Long, val title: String, val content: String, val weekDay: WeekDay, val order: Int)
